pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bootcs"
version = "2.0.0"
description = "BootCS CLI - Check and submit your code, based on check50/lib50"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.9"
authors = [{ name = "BootCS Team", email = "team@bootcs.dev" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]
dependencies = [
    "pexpect>=4.8",
    "termcolor>=2.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "attrs>=23.0",
    "jellyfish>=1.0",
    "packaging>=23.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "black>=23.0", "ruff>=0.1"]

[project.scripts]
bootcs-cli = "bootcs.__main__:main"

[project.urls]
Homepage = "https://bootcs.dev"
Repository = "https://github.com/bootcs-dev/bootcs-cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["bootcs*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
